Extract population analysis into dedicated src/populations/ module
Splits population analysis and multipole-moment evaluation out of the
scf/ and integrals/ modules into a standalone src/populations/ tree so
the analyses live next to one another and can grow independently of the
SCF and integral engines.

- Move scf/population.{cpp,h} into populations/, splitting the monolith
  into bond-order.cpp, lodwin.cpp, mulliken.cpp, population.h, and
  population_detail.h
- Move the multipole-matrix and multipole-moment routines out of
  integrals/os.{cpp,h} (_os_1d_moments, _compute_multipole_matrices,
  _compute_multipole_moments) into populations/multipole.{cpp,h}
- Update src/driver.cpp, src/dft/driver.cpp, and src/dft/main.cpp to
  include the new headers
- Update CMakeLists.txt with a POPULATIONS_SRC glob, link it into both
  hartree-fock and planck-dft, and point the multipole_integrals and
  population_analysis test targets at the new sources
- Update test includes to use populations/ headers
