+++ src/dft/main.cpp
@@ -7,9 +7,9 @@
 #include <string>
 
 #include "driver.h"
-#include "integrals/os.h"
 #include "io/io.h"
 #include "io/logging.h"
+#include "populations/multipole.h"
 
 using SystemClock = std::chrono::system_clock;
