+++ src/driver.cpp
@@ -18,10 +18,11 @@
 #include "io/logging.h"
 #include "lookup/elements.h"
 #include "opt/geomopt.h"
+#include "populations/multipole.h"
+#include "populations/population.h"
 #include "post_hf/casscf.h"
 #include "post_hf/cc.h"
 #include "post_hf/mp2.h"
-#include "scf/population.h"
 #include "scf/scf.h"
 #include "symmetry/integral_symmetry.h"
 #include "symmetry/mo_symmetry.h"
